stubbed out / completed 85% of logic for city name
